refactor(shared): introduce DateTimeUtility for common date operations

+++ shared/src/commonMain/kotlin/de/berlinskylarks/shared/data/model/Game.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class Game(
-    override var id: Int,
+    override var id: Long,
     @SerialName("match_id")
     var matchID: String,
     @SerialName("planned_innings")
