+++ shared/src/commonMain/kotlin/de/berlinskylarks/shared/data/model/JSONDataObject.kt
@@ -1,5 +1,5 @@
 package de.berlinskylarks.shared.data.model
 
 interface JSONDataObject {
-    var id: Int
+    var id: Long
 }
